feat: Add modal root element to index.html and update font in Layout component
This commit adds a new div element with the id "modal-root" to the index.html file in the client directory. This element will be used to render modal components in the application.

Additionally, the font in the Layout component is updated to use the "Raleway" font family in addition to the existing "Roboto" font family. This change is made by adding the "font-raleway" class to the main element in the Layout component.

These changes improve the functionality and styling of the application.